refactor: move grpc wrapper
